refactor: move statusReport implementations to /services/base.go
