+++ backend/pkg/exporter/modules/network_liveness_test.go
@@ -76,7 +76,6 @@ func TestNetworkLivenessExport(t *testing.T) {
 		ctx:    ctx,
 		cache:  tieredCache,
 	}
-	StatusReporter = stubStatusReporter{}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
